REST Controller Error Handling

+++ 03-03-rest-error-handling/src/main/java/com/linor/singer/controller/SingerController.java
@@ -15,6 +15,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.linor.singer.dao.SingerDao;
 import com.linor.singer.domain.Singer;
+import com.linor.singer.exception.BizException;
+import com.linor.singer.exception.DataAccessException;
 import com.linor.singer.exception.ResourceNotFoundException;
 
 @RestController
@@ -33,11 +35,14 @@ public class SingerController {
 		Singer singer = singerDao.findById(id);
 		if(singer == null)
 			throw new ResourceNotFoundException();
+		
 		return singer;
 	}
 	
 	@PostMapping
 	public void addSinger(@RequestBody Singer singer) {
+		if(null != singer.getId())
+			throw new BizException("비즈니스 오류 발생");
 		singerDao.insertWithAlbum(singer);
 	}
 	
@@ -47,7 +52,10 @@ public class SingerController {
 	}
 	
 	@DeleteMapping(value = "/{id}")
-	public void deleteSinger(@PathVariable("id") int id) {
+	public void deleteSinger(@PathVariable("id") int id) throws Exception{
+		Singer singer = singerDao.findById(id);
+		if (singer == null)
+			throw new DataAccessException("데이타가 존재하지 않음");
 		singerDao.delete(id);
 	}
 }
